devotv-automation/libraries/requirements.txt
Faker
pep8
pycodestyle
pytest
requests
RESTinstance
robotframework
robotframework-faker
robotframework-requests
robotframework-seleniumlibrary
selenium
pyautogui
xmltodict
mysql-connector-python
openapi-schema-validator

